+++ src/components/TeacherViews/SpecialistProgrammeListing/SpecialistProgrammeListing.fixture.ts
@@ -1,0 +1,27 @@
+export const specialistProgrammeListing = {
+  subjectSlug: "communication-and-language",
+  subjectTitle: "Communication and language",
+  programmes: [
+    {
+      programmeSlug: "communication-and-language-ks1",
+      developmentStageSlug: "ks1",
+      developmentStageTitle: "KS1",
+      unitCount: 10,
+      lessonCount: 100,
+    },
+    {
+      programmeSlug: "communication-and-language-ks2",
+      developmentStageSlug: "ks2",
+      developmentStageTitle: "KS2",
+      unitCount: 10,
+      lessonCount: 100,
+    },
+    {
+      programmeSlug: "communication-and-language-ks3",
+      developmentStageSlug: "ks3",
+      developmentStageTitle: "KS3",
+      unitCount: 10,
+      lessonCount: 100,
+    },
+  ],
+};