+++ src/components/TeacherViews/SpecialistProgrammeListing/SpecialistProgrammeListing.view.stories.tsx
@@ -1,0 +1,16 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import Component from "./SpecialistProgrammeListing.view";
+import { specialistProgrammeListing } from "./SpecialistProgrammeListing.fixture";
+const meta: Meta<typeof Component> = {
+  component: Component,
+};
+
+export default meta;
+type Story = StoryObj<typeof Component>;
+
+export const SpecialistProgrammeListingPage: Story = {
+  args: {
+    ...specialistProgrammeListing,
+  },
+};