+++ src/node-lib/curriculum-api-2023/index.ts
@@ -22,6 +22,10 @@ import lessonShareQuery from "./queries/lessonShare/lessonShare.query";
 import specialistSubjectListingQuery from "./queries/specialistSubjectListing/specialistSubjectListing.query";
 import { pupilLessonOverviewCanonicalQuery } from "./queries/pupilLessonOverviewCanonical/pupilLessonOverviewCanonical.query";
 import specialistUnitListingQuery from "./queries/specialistUnitListing/specialistUnitListing.query";
+import specialistProgrammeListingQuery from "./queries/specialistProgrammeListing/specialistProgrammeListing.query";
+import specialistLessonListingQuery from "./queries/specialistLessonListing/specialistLessonListing.query";
+import { specialistLessonDownloadQuery } from "./queries/specialistLessonDownload/specialistLessonDownload.query";
+import { specialistLessonShareQuery } from "./queries/specialistLessonShare/specialistLessonShare.query";
 
 export const keyStageSchema = z.object({
   slug: z.string(),
@@ -136,6 +140,10 @@ const curriculumApi2023 = {
   unitListing: unitListingQuery(sdk),
   specialistSubjectListing: specialistSubjectListingQuery(sdk),
   specialistUnitListing: specialistUnitListingQuery(sdk),
+  specialistProgrammeListing: specialistProgrammeListingQuery(sdk),
+  specialistLessonListing: specialistLessonListingQuery(sdk),
+  specialistLessonDownloads: specialistLessonDownloadQuery(sdk),
+  specialistLessonShare: specialistLessonShareQuery(sdk),
 };
 
 export type CurriculumApi = typeof curriculumApi2023;