+++ src/node-lib/curriculum-api/index.ts
@@ -570,7 +570,6 @@ const curriculumApi = {
   lessonListing: async (...args: Parameters<typeof sdk.lessonListing>) => {
     const res = await sdk.lessonListing(...argsRemoveLegacySlugSuffix(args));
     const { units = [], lessons = [] } = transformMVCase(res);
-
     const unit = getFirstResultOrWarnOrFail()({
       results: units,
     });